Fixed errors in homogeneity calculation
Fixed a bug in the homogeneity calculation that resulted in consistent underestimation of the homogeneity.
The issue was tested on an ideal sinusoid surface, resulting in an H of only ~0.9, when it should be 1.0.
This could be traced back to the parameter Ssk (skewness), which is close to zero when the material is distributed
equally around the the mean. For an ideal surface sine surface, this parameter should have been zero, but was not
due to the shortcomings of floating point arithmetic. Since all values were close to zero, the gini coefficent
was quite large and the homogeneity therefore low, bringing down the total homogeneity averaged over all considered
roughness parameters.