Removed w/h as parameters for Surface
width_um and height_um were removed as initialization parameters for Surface objects,
since they represent redundant information. Instead, they are calculated from the shape
of the data as well as the stepsize. Moreover, lru_cache was added to some methods. In order
to clear the cache whenever the data is changed inplace, a new method Surface._set_data was
introduced to handle both the cache clearing as well as the recalculation of the width and height.