Added functional volume parameters
Added Vmp, Vmc, Vvv, Vvc parameter calculations. For this purpose, the calculation of all functional parameters
was moved to a new class: AbbottFirestoneCurve. This class now handles the calculation of the equivalence line
all all associated values. Thereby, these values can be resused for mulÃtiple parameter calculations as well as
for plotting.