Fixed bug with surface area calculation
Fixed a bug that caused negative Sdr for flat and nearly flat surfaces. This bug was caused by a surface area
smaller than the projected area that originated from the variable 'total_area' in '_surface_area_iso_cy' being
a float and not a double. Same error was fixed in the gwyddion version of the area calculation.