Added spatial parameters Sal and Str
Added the calculation of spatial parameters by means of autocorrelation.
Added a class that holds the data and calculations for the autocorrelation
function and derives Sal and Str parameters. Added tests for Sal and Str.
Updated changelog. Updated tests for Sdr, since in the prior commit, a bug in
the Sdr calculation was fixed that resulted in slightly off values, which failed
the old tests.

Note: The test for the spatial period didn't pass by a slight margin. Unclear why.
Should be investigated.