Fixed filter method
Fixed errors in filtering method where correct frequencies were only filtered in x, not in y
direction because the different sampling length and consequently the different frequency resolutions
were not considered. Moreover, the rectangular extent in x-y was replaced by an ellipsis.
Now, misaligned textures are also filtered at the correct frequencies.
Nonetheless, this method still uses zeroing of the bins of the DFT, which still results in artifacts
from the application of a rectangular windowing function. This has to be dealt with in the future
by implementing correct Gaussian or Spline filtering. Lastly, endeffects are not managed at the moment.