Added size property to Surface
Added a size property to Surface that replaces the call to Surface.data.shape,
since it is better not to expose the internals. Instead, Surface.size returns
a namedtuple of type size, which behaves like the previous tuple, but with the
added functionality that the size in x and y can be accessed by size.x and size.y.
This provides some clarity since the first dimension in the tuple is y and the
second is x, which might be counterintuitive. Now, all code accesses these dimensions
by Surface.size.x and Surface.size.y instead of Surface.data.shape[1] and
Surface.data.shape[0].