Reworked Batch processing
Batch now doesn't load the surfaces on initialization. Instead, operations and parameters
are registered and are only applied to the surfaces upon calling of the Batch.execute method.
The advantage is firstly, that not all surface files have to be loaded into memory at once,
which could be a´n issue for larger batch size. Secondly, this is the basis for sharing the
processing load among different cores using multiprocesing. Right now, multiprocessing doesn't
work yet, this remains to be solved. Moreover, a method for calculation of all parameters at
once as well as the merging of the data into a dataframe needs to be reimplemented.