Reworked functional parameter calculation
The AbbottFirestoneCurve is now calculated from all available datapoints without binning. This
means that the arrays for height and material ratio are now both unevenly spaced. However, the benefit
is that we increase the accuracy and since we are using linear interpolators anyway, we can still evaluate
the function at any point including at regularly spaced intervals.