Reworked sinusoid fit
Sinusoid is now a class that can be constructued from sinusoid parameters and evaluated
using __call__ method. It has a class method that allows to construct it by fitting and
a method which calculates the position of the first extremum. The methods of Surface that
rely on the sinusoid have been adapted to the changes. Moreover, Surface.depth now starts
to compute the depth intervals starting from the position of the first extremum, which was
not guaranteed before. However, this seems to cause some issues with the depth calculation
which need to be fixed.