Switch from Multiprocessing Pool to Threadpool for Batch execution
This implementation was switched from a true multiprocessing pool to a thread pool. The reason for the change is that multiprocessing relies on pickling, which causes multiple issues when using Jupyter Notebooks. Also, if the batch.execute method is not called in a main guard, an infinite spawning of child processes may occur. These issues are avoided when using threads while maintaining most of the speedup, since most numpy-based computations release the GIL anyway.