Refactoring of autocorrelation function
Correction and refactoring of the ACF. The values for Sal and Str remain the same. The calculation of the ACF should be more efficient and yield correct values. Moreover, it is now separate from the calculation of the values necessary for Sal and Str. Lastly, the ACF has received plotting functionality.