Refactored file reading and writing
File reader and writer functions now register with a new FileHandler class
with their file suffix and file magic using a decorator. The FileHandler
offers read and write methods, which dispatch the operation to the correct
reader or writer. The reader is first inferred from the suffix. However, if
the reader for the correct suffix raises an Exception, the FileHandler checks
the other readers for a matching file magic, in case the file has the wrong
file extension. If a reader is found based on the file magic, the file is read
and a warning is issued.
The loader.py module was removed since it is now obsolete. The FileHandler is
directly imported in the __init__.py of the file package. In order for the
reader and writer functions to register with the FileHandler class through its
decorator class method, the respective modules must be imported once. This is done
by iterating over all available module files and importing them in the __init__.py.
Thereby, new file formats do not have to explicitly register in the __init__.py.
This seems more elegent but could also be confusing, therefore I might reverse that
implementation in the future.