Refactored file readers and writers to work with buffers
File readers and writers now take an opened file object or buffer as
first argument instead of filepath. The FileHandler class is now responsible
for opening the files or passing on the buffers. File formats can now be
specified in Surface.save and Surface.load methods to write to and read from
buffers. Added tests for reading and writing operations with buffers.